Eliminate circular dependency in validation rules
`validation/validate.js` depends on all of the rules, and all of the
rules depend on `ValidationContext`, which was previously exported by
`validation/validate.js`.

This moves `ValidationContext` out of `validate.js` to resolve this.

Circular dependencies are fine, but they tend to slow down flow, since
it moves more work to the merging phase, and the merging phase isn't as
fast as the inference phase. As a result, this should make flow
marginally faster.

This does have a minor breaking change: `graphql/validatation/validate`
no longer exports `ValidationContext`, but `graphql/validation` still
does. I can add an export to `validate.js` to preserve that behavior,
but it's a minor change to a non-public API, and graphql-js has broken
these things in the past.
