+++ src/utilities/isValidLiteralValue.js
@@ -15,7 +15,7 @@ import { visit, visitWithTypeInfo } from '../language/visitor';
 import type { GraphQLInputType } from '../type/definition';
 import { GraphQLSchema } from '../type/schema';
 import { ValuesOfCorrectType } from '../validation/rules/ValuesOfCorrectType';
-import { ValidationContext } from '../validation/validate';
+import ValidationContext from '../validation/ValidationContext';
 
 /**
  * Utility which determines if a value literal node is valid for an input type.
