+++ src/validation/rules/FragmentsOnCompositeTypes.js
@@ -7,7 +7,7 @@
  * @flow strict
  */
 
-import type { ValidationContext } from '../index';
+import type ValidationContext from '../ValidationContext';
 import { GraphQLError } from '../../error';
 import { print } from '../../language/printer';
 import type { ASTVisitor } from '../../language/visitor';
