+++ src/validation/rules/OverlappingFieldsCanBeMerged.js
@@ -7,7 +7,7 @@
  * @flow strict
  */
 
-import type { ValidationContext } from '../index';
+import type ValidationContext from '../ValidationContext';
 import { GraphQLError } from '../../error';
 import find from '../../jsutils/find';
 import type { ObjMap } from '../../jsutils/ObjMap';
