+++ src/validation/rules/SingleFieldSubscriptions.js
@@ -7,7 +7,7 @@
  * @flow strict
  */
 
-import type { ValidationContext } from '../index';
+import type ValidationContext from '../ValidationContext';
 import { GraphQLError } from '../../error';
 import type { OperationDefinitionNode } from '../../language/ast';
 import type { ASTVisitor } from '../../language/visitor';
