+++ src/validation/rules/UniqueVariableNames.js
@@ -7,7 +7,7 @@
  * @flow strict
  */
 
-import type { ValidationContext } from '../index';
+import type ValidationContext from '../ValidationContext';
 import type { VariableDefinitionNode } from '../../language/ast';
 import { GraphQLError } from '../../error';
 import type { ASTVisitor } from '../../language/visitor';
