+++ src/validation/rules/ValuesOfCorrectType.js
@@ -7,7 +7,7 @@
  * @flow strict
  */
 
-import type { ValidationContext } from '../index';
+import type ValidationContext from '../ValidationContext';
 import { GraphQLError } from '../../error';
 import type { ValueNode } from '../../language/ast';
 import { print } from '../../language/printer';
